Changed game timers to uint32_t

+++ OpenGL/src/controller/KeyboardController.hpp
@@ -28,5 +28,5 @@ private:
   bool dab;
 
   bool spaceIsPressed;
-  float lastSpacePress;
+  uint32_t lastSpacePress;
 };
